fixed nulls

+++ src/dev/scroopid/crafexEngine/graphics/GraphicsUtil.java
@@ -54,7 +54,8 @@ public class GraphicsUtil {
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-		int scale = Crafex.WINDOW_DIMENTIONS.getY() / Crafex.WINDOW_DEFAULT.getY();
+		
+		float scale = Crafex.WINDOW_DIMENTIONS.getY() / Crafex.WINDOW_DEFAULT.getY();
 
 		if (temp != null) {
 			data = scaleImage(temp, scale);
